Create README.md
INTEGRATED VESSELS NETWORK LTD is limited liability company established in October 2022 under the law of the federal Republic of Nigeria with a CAC NO: 123456. Its provides many services like the supply of consumables to offshore platforms, supply of ship stores and spare parts to shipping companies and majorly creating synergy between ship owners and charterers to seamless hiring of vessels and marines equipment.